Added ValueProvider interface.
This interface is responsible for creating a new Value object.
Its method 'newValue' needs a quantity and a ValueDomain as parameters.
ValueDomain objects will call this method whenever they need to return a value object when an instance of a targeted class.
They will use the target instance as the quantity parameter and itself as the domain parameter